+++ app/api/build.gradle.kts
@@ -9,6 +9,6 @@ tasks.getByName("jar") {
 dependencies {
     implementation("org.springframework.boot:spring-boot-starter-web")
     developmentOnly("org.springframework.boot:spring-boot-devtools")
-    implementation(project(":app:dominio"))
-    implementation(project(":app:aplicacao"))
+    implementation(project(":app:domain"))
+    implementation(project(":app:application"))
 }
