+++ app/api/src/main/kotlin/br/com/fiap/gerenciamentopedidos/api/config/AppBeansConfig.kt
@@ -1,20 +1,20 @@
 package br.com.fiap.gerenciamentopedidos.api.config
 
-import br.com.fiap.gerenciamentopedidos.aplicacao.casosdeuso.CadastrarPedidoCasoDeUso
-import br.com.fiap.gerenciamentopedidos.dominio.interfaces.services.PedidoServico
-import br.com.fiap.gerenciamentopedidos.dominio.servicos.PedidoServicoImpl
+import br.com.fiap.gerenciamentopedidos.application.usecases.CadastrarPedidoUseCase
+import br.com.fiap.gerenciamentopedidos.domain.interfaces.services.PedidoService
+import br.com.fiap.gerenciamentopedidos.domain.services.PedidoServiceImpl
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 
 @Configuration
 class AppBeansConfig {
     @Bean
-    fun pedidoServico(): PedidoServico {
-        return PedidoServicoImpl()
+    fun pedidoServico(): PedidoService {
+        return PedidoServiceImpl()
     }
 
     @Bean
-    fun cadastrarPedidoCasoDeUso(pedidoServico: PedidoServico): CadastrarPedidoCasoDeUso {
-        return CadastrarPedidoCasoDeUso(pedidoServico)
+    fun cadastrarPedidoCasoDeUso(pedidoService: PedidoService): CadastrarPedidoUseCase {
+        return CadastrarPedidoUseCase(pedidoService)
     }
 }
