--- app/api/src/main/kotlin/br/com/fiap/gerenciamentopedidos/api/controladores/PingController.kt
@@ -1,15 +1,0 @@
-package br.com.fiap.gerenciamentopedidos.api.controladores
-
-import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.RequestMapping
-import org.springframework.web.bind.annotation.RestController
-
-@RestController
-@RequestMapping("/ping")
-class PingController {
-
-    @GetMapping
-    fun ping(): String {
-        return "Pong"
-    }
-}
