+++ app/api/src/main/kotlin/br/com/fiap/gerenciamentopedidos/api/controllers/PingController.kt
@@ -1,0 +1,15 @@
+package br.com.fiap.gerenciamentopedidos.api.controllers
+
+import org.springframework.web.bind.annotation.GetMapping
+import org.springframework.web.bind.annotation.RequestMapping
+import org.springframework.web.bind.annotation.RestController
+
+@RestController
+@RequestMapping("/ping")
+class PingController {
+
+    @GetMapping
+    fun ping(): String {
+        return "Pong"
+    }
+}
