--- app/aplicacao/src/main/kotlin/br/com/fiap/gerenciamentopedidos/aplicacao/interfaces/CasoDeUso.kt
@@ -1,4 +1,0 @@
-package br.com.fiap.gerenciamentopedidos.aplicacao.interfaces
-
-interface CasoDeUso {
-}
