+++ app/application/src/main/kotlin/br/com/fiap/gerenciamentopedidos/application/interfaces/UseCase.kt
@@ -1,0 +1,4 @@
+package br.com.fiap.gerenciamentopedidos.application.interfaces
+
+interface UseCase {
+}
