+++ app/domain/src/main/kotlin/br/com/fiap/gerenciamentopedidos/domain/interfaces/repositories/PedidoRepository.kt
@@ -1,0 +1,4 @@
+package br.com.fiap.gerenciamentopedidos.domain.interfaces.repositories
+
+interface PedidoRepository {
+}
