+++ app/domain/src/main/kotlin/br/com/fiap/gerenciamentopedidos/domain/models/Pedido.kt
@@ -1,0 +1,4 @@
+package br.com.fiap.gerenciamentopedidos.domain.models
+
+class Pedido {
+}
