+++ app/domain/src/main/kotlin/br/com/fiap/gerenciamentopedidos/domain/services/PedidoServiceImpl.kt
@@ -1,0 +1,6 @@
+package br.com.fiap.gerenciamentopedidos.domain.services
+
+import br.com.fiap.gerenciamentopedidos.domain.interfaces.services.PedidoService
+
+class PedidoServiceImpl : PedidoService {
+}
