--- app/dominio/src/main/kotlin/br/com/fiap/gerenciamentopedidos/dominio/interfaces/repositorios/PedidoRepositorio.kt
@@ -1,4 +1,0 @@
-package br.com.fiap.gerenciamentopedidos.dominio.interfaces.repositorios
-
-interface PedidoRepositorio {
-}
