--- app/dominio/src/main/kotlin/br/com/fiap/gerenciamentopedidos/dominio/interfaces/services/PedidoServico.kt
@@ -1,4 +1,0 @@
-package br.com.fiap.gerenciamentopedidos.dominio.interfaces.services
-
-interface PedidoServico {
-}
