--- app/dominio/src/main/kotlin/br/com/fiap/gerenciamentopedidos/dominio/modelos/Pedido.kt
@@ -1,4 +1,0 @@
-package br.com.fiap.gerenciamentopedidos.dominio.modelos
-
-class Pedido {
-}
