--- app/dominio/src/main/kotlin/br/com/fiap/gerenciamentopedidos/dominio/servicos/PedidoServicoImpl.kt
@@ -1,6 +1,0 @@
-package br.com.fiap.gerenciamentopedidos.dominio.servicos
-
-import br.com.fiap.gerenciamentopedidos.dominio.interfaces.services.PedidoServico
-
-class PedidoServicoImpl : PedidoServico {
-}
