--- app/infraestrutura/build.gradle.kts
@@ -1,3 +1,0 @@
-dependencies {
-    implementation(project(":app:dominio"))
-}
