--- app/infraestrutura/src/main/kotlin/br/com/fiap/gerenciamentopedidos/infraestrutura/repositorios/PedidoRepositorioImpl.kt
@@ -1,6 +1,0 @@
-package br.com.fiap.gerenciamentopedidos.infraestrutura.repositorios
-
-import br.com.fiap.gerenciamentopedidos.dominio.interfaces.repositorios.PedidoRepositorio
-
-class PedidoRepositorioImpl : PedidoRepositorio {
-}
