+++ app/infrastructure/build.gradle.kts
@@ -1,0 +1,3 @@
+dependencies {
+    implementation(project(":app:domain"))
+}
