+++ app/infrastructure/src/main/kotlin/br/com/fiap/gerenciamentopedidos/infrastructure/repositories/PedidoRepositoryImpl.kt
@@ -1,0 +1,6 @@
+package br.com.fiap.gerenciamentopedidos.infrastructure.repositories
+
+import br.com.fiap.gerenciamentopedidos.domain.interfaces.repositories.PedidoRepository
+
+class PedidoRepositoryImpl : PedidoRepository {
+}
